pyproject.toml
[project]
name = "dataverse-to-mailchimp"
version = "0.1.0"
description = ""
authors = [
    {name = "Ed Shortt",email = "ed.shortt@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "msal (>=1.32.3,<2.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
